requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
python-multipart>=0.0.20
openai>=1.59.3,<2.0.0
tavily-python>=0.5.0
resend>=2.4.0
requests>=2.32.3
beautifulsoup4>=4.12.3
pydantic>=2.11.0
python-dotenv>=1.0.1
trafilatura>=2.0.0
tenacity>=9.0.0
